+++ src/content/lib/request-processor.compat.js
@@ -23,7 +23,7 @@
 
 Cu.import("resource://gre/modules/AddonManager.jsm");
 
-Cu.import("chrome://requestpolicy/content/lib/script-loader.jsm");
+Cu.import("chrome://rpcontinued/content/lib/script-loader.jsm");
 ScriptLoader.importModules([
   "lib/logger",
   "lib/utils",